use uint64_t for sfxsize section

+++ php_micro_fileinfo.h
@@ -25,8 +25,8 @@ limitations under the License.
 
 #include "php.h"
 
-uint32_t micro_get_sfxsize(void);
-uint32_t micro_get_sfxsize_limit(void);
+size_t micro_get_sfxsize(void);
+size_t micro_get_sfxsize_limit(void);
 
 #ifdef PHP_WIN32
 /*
